requirements.txt
pandas~=2.2.2
pip~=25.3
typing_extensions~=4.12.2
attrs~=23.2.0
filelock~=3.9.0
PyYAML~=6.0
numpy~=1.24.1
joblib~=1.4.2
scikit-learn~=1.5.1
xgboost~=3.0.5